% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-fa-like.R
\name{as_fa_like}
\alias{as_fa_like}
\alias{as_fa_like.list}
\title{Coerce an object to a factor analysis like factorization}
\usage{
as_fa_like(x, ...)

\method{as_fa_like}{list}(x, ...)
}
\arguments{
\item{x}{Object to coerce}

\item{...}{Ignored.}
}
\value{
Object as \code{\link[=svd_like]{svd_like()}} object.
}
\description{
Coerce an object to a factor analysis like factorization
}
