% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim_and_class}
\alias{dim_and_class}
\title{Description array size and type}
\usage{
dim_and_class(x)
}
\arguments{
\item{x}{Matrix or vector}
}
\value{
Description as character vector
}
\description{
Description array size and type
}
