% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{masked_approximation}
\alias{masked_approximation}
\title{TODO}
\usage{
masked_approximation(object, mask)
}
\arguments{
\item{object}{}

\item{mask}{}
}
\description{
TODO: this needs to become a generic so it
can also work for factor analysis objects
}
