% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-svd-like.R
\name{as_svd_like}
\alias{as_svd_like}
\title{Coerce an object to LRMF class}
\usage{
as_svd_like(x, ...)
}
\arguments{
\item{x}{Object to coerce}

\item{...}{Ignored.}
}
\value{

}
\description{
Coerce an object to LRMF class
}
