% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fa_like}
\alias{predict.fa_like}
\title{Title}
\usage{
\method{predict}{fa_like}(object, X = NULL, ...)
}
\arguments{
\item{object}{TODO}

\item{X}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Title
}
