% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{predict.LRMF}
\alias{predict.LRMF}
\alias{residuals.LRMF}
\alias{fitted.LRMF}
\title{TODO}
\usage{
\method{predict}{LRMF}(object, X = NULL, ...)

\method{residuals}{LRMF}(object, X, ...)

\method{fitted}{LRMF}(object, X = NULL, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}

\item{mask}{TODO}
}
\value{

}
\description{
TODO
}
\examples{

X <- as.matrix(trees)
tree_mf <- as_mf(svd(X))

predict(tree_mf)
fitted(tree_mf)
resid(tree_mf, X)

}
