% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.svd_like}
\alias{predict.svd_like}
\title{TODO}
\usage{
\method{predict}{svd_like}(object, X = NULL, ...)
}
\arguments{
\item{object}{TODO}

\item{X}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{

X <- as.matrix(trees)
tree_mf <- as_svd_like(svd(X))

predict(tree_mf)
resid(tree_mf, X)


}
