% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRMF3-package.R
\docType{package}
\name{LRMF3-package}
\alias{LRMF3}
\alias{LRMF3-package}
\title{LRMF3: Low Rank Matrix Factorization S3 Objects}
\description{
Provides S3 classes to represent low rank matrix decompositions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RoheLab/LRMF3}
  \item Report bugs at \url{https://github.com/RoheLab/LRMF3/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})

}
\keyword{internal}
