% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mf.R, R/object.R
\name{as_mf}
\alias{as_mf}
\alias{mf}
\title{Coerce an object to LRMF class}
\usage{
as_mf(x, ...)

mf(u, d, v, subclasses = NULL, ...)
}
\arguments{
\item{x}{Object to coerce}

\item{...}{Optional additional items to pass to the constructor.}

\item{u}{The "left singular-ish" vectors.}

\item{d}{The "singular-ish" values.}

\item{v}{The "right singular-ish" vectors.}

\item{subclasses}{A character vector of subclasses.
Optional, defaults to \code{NULL}.}
}
\value{

}
\description{
A low rank matrix factorization of a matrix \code{X} is
parameterized by \verb{X ~= u \%*\% diag(d) \%*\% t(v)}.
}
\examples{

s <- svd(as.matrix(trees))

# using the constructor directly
mf(s$u, s$d, s$v)

# coercing svd-like lists
as_mf(s)

}
